+++ Dental_Clinic_Register_BackEnd/src/main/java/edu/ubb/licenta/pgim2289/spring/service/BraceComponentService.java
@@ -5,7 +5,7 @@ import edu.ubb.licenta.pgim2289.spring.dto.BraceComponentDTO;
 import java.util.List;
 
 public interface BraceComponentService {
-    List<BraceComponentDTO> syncBraceComponents(Integer treatmentPlanId, List<BraceComponentDTO> requestList);
+    List<BraceComponentDTO> syncBraceComponents(Long treatmentPlanId, List<BraceComponentDTO> requestList);
 
-    List<BraceComponentDTO> getBraceComponentsByTreatmentPlanId(Integer treatmentPlanId);
+    List<BraceComponentDTO> getBraceComponentsByTreatmentPlanId(Long treatmentPlanId);
 }
